feature[ats_ATS19845]: Fix build errors due to branch_id to long

+++ plugins/org.eclipse.osee.framework.core.message/src/org/eclipse/osee/framework/core/message/internal/translation/BranchChangeArchivedStateRequestTranslator.java
@@ -27,7 +27,7 @@ public final class BranchChangeArchivedStateRequestTranslator implements ITransl
 
    @Override
    public ChangeBranchArchiveStateRequest convert(PropertyStore propertyStore) {
-      int branchId = propertyStore.getInt(Entry.BRANCH_ID.name());
+      long branchId = propertyStore.getLong(Entry.BRANCH_ID.name());
       BranchArchivedState state = BranchArchivedState.valueOf(propertyStore.get(Entry.BRANCH_ARCHIVED_STATE.name()));
       ChangeBranchArchiveStateRequest data = new ChangeBranchArchiveStateRequest(branchId, state);
       return data;
